+++ example_impl/example_testsuite/example_testsuite.go
@@ -1,16 +1,32 @@
+/*
+ * Copyright (c) 2020 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
 package example_testsuite
 
 import (
 	"github.com/kurtosis-tech/kurtosis-go/lib/testsuite"
 )
 
-// TODO example of parameterizing your image
-type ExampleTestsuite struct {}
+type ExampleTestsuite struct {
+	serviceImage string
+}
 
-func (e ExampleTestsuite) GetTests() map[string]testsuite.Test {
+func NewExampleTestsuite(serviceImage string) *ExampleTestsuite {
+	return &ExampleTestsuite{serviceImage: serviceImage}
+}
+
+
+func (suite ExampleTestsuite) GetTests() map[string]testsuite.Test {
 	return map[string]testsuite.Test{
-		"exampleTest1": ExampleTest1{},
-		"exampleTest2": ExampleTest2{},
+		"singleNodeExampleTest": SingleNodeExampleTest{ServiceImage: suite.serviceImage},
+		"fixedSizeExampleTest": FixedSizeExampleTest{ServiceImage: suite.serviceImage},
 	}
 }
+
+func (suite ExampleTestsuite) GetNetworkWidthBits() uint32 {
+	return 8
+}
+