+++ CHANGELOG.md
@@ -1,5 +1,15 @@
-## TBD
+## 1.0.0
 * Created example test suite to validate that the client library work
 * Bugfix in volume-writing location, and force pretty formatting on written logs
 * Made the existing test actually query the node it created
 * Added another test to demonstrate an initial network setup
+* Adding copyright headers
+* Renamed tests to have more descriptive names
+* When asked about test suite data, send back a JSON of test suite metadata (rather than just a list of test names)
+* Made log level configurable
+* Add CircleCI
+* Upgraded example Go implementation to show the use of custom environment variables
+* Build a Docker image on each merge to the develop branch
+* Accept a new Docker parameter, `SERVICES_RELATIVE_DIRPATH`, for the location (relative to the suite execution volume root) where file IO for the services created during test execution
+* Consolidate all the scripts into `build_and_run.sh` which will actually run the test suite for testing purposes
+* Switch to `master` release track from Kurtosis core